// app/types/AuditoriumSchedule.ts
export type AuditoriumSchedule = {
  // idSistema: string;
  dataInicio: string;
  dataFim: string;
  // complexo: string;
  // idUnidade: string;
  sala: string;
  versao: string;
  idioma: string;
  horarios: string[];
  // idHtticket: string;
  idERP: string;
  idFilme: string;
  filmeUrl: string;
};